qpid-1277: fixed header conversion issues + added test
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@701107 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/dotnet/client-010/client/transport/codec/MSDecoder.cs
@@ -76,7 +76,7 @@ namespace org.apache.qpid.transport.codec
 
         public override long readUint64()
 		{
-		    return (long) ByteEncoder.GetBigEndian((Double) reader.ReadInt64());            
+		    return (long) ByteEncoder.GetBigEndian(reader.ReadInt64());            
 		}
 	}
 }
